(PC-19021)[API] refactor: move amplitude tracker function to analytics folder
